Adjusted spam checks
This commit adjusts the order of spam checks so that character spam,
then capitalized letters, and then capitalized words are checked. This
commit also ensures that a player doesn't get a message for more than
one type of spam.

+++ src/main/java/com/haleykell/spamfilter/commands/ChangeSpamFilterCommand.java
@@ -20,8 +20,8 @@ public class ChangeSpamFilterCommand implements CommandExecutor {
         if (strings.length == 0) {
             commandSender.sendMessage(Lang.SPAMFILTER_TITLE + "Current spam limits are");
             commandSender.sendMessage(Lang.SPAMFILTER_TITLE + "Character Spam: " + config.getInt("limit"));
-            commandSender.sendMessage(Lang.SPAMFILTER_TITLE + "Capitalized Words: " + config.getInt("capsWords"));
-            commandSender.sendMessage(Lang.SPAMFILTER_TITLE + "Capitalized Letters: " + config.getInt("caps"));
+            commandSender.sendMessage(Lang.SPAMFILTER_TITLE + "Capitalized Words: " + config.getDouble("capsWords"));
+            commandSender.sendMessage(Lang.SPAMFILTER_TITLE + "Capitalized Letters: " + config.getDouble("caps"));
             return true;
         }
         else if (strings.length > 2 || strings.length == 1) {
